+++ app/src/main/AndroidManifest.xml
@@ -12,10 +12,10 @@
         android:theme="@style/Theme.Shopping"
         tools:targetApi="31">
         <activity
-            android:name=".cart.CartActivity"
+            android:name=".view.cart.CartActivity"
             android:exported="false" />
         <activity
-            android:name=".ProductListActivity"
+            android:name=".view.productlist.ProductListActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
