+++ app/src/main/res/layout/activity_cart.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".cart.CartActivity">
+    tools:context=".view.cart.CartActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
