+++ app/src/main/res/layout/activity_product_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProductListActivity">
+    tools:context=".view.productlist.ProductListActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/grid_products"
